+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Connection closure on remote side when TLS is enabled is now handled (#48).
 - `alt-char` bindings implemented to switching between tabs.
 - Fixed some bugs in `join` command used in `auto_cmds` (#49, #38).
+- Tabs can now be moved left/right with `alt-left/right` keys (#52).
+- Input field cursor location now preserved after resize.
+- `TOPIC` messages are now handled (#58).
+- `RPL_AWAY` is now handled (#55). Away message is shown in user tab.
 
 # 2017/11/12: 0.3.0